% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edh.R
\name{ehd}
\alias{ehd}
\title{Extract strings from a single header}
\usage{
ehd(header_vec)
}
\arguments{
\item{header_vec}{wave header string}
}
\value{
vector holding header data (recording time, time zone, audiomoth serial number, gain, voltage, temperature)
}
\description{
Extract strings from a single header
}
\examples{
test_string <- "Recorded at 12:21:00 10/02/2021 (UTC) by AudioMoth 24E144085F256717 at medium gain setting while battery state was 3.9V and temperature was -2.6C. Amplitude threshold was 160."
ehd(test_string)
}
