% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_audiomoth_files.R
\name{rename_audiomoth_files}
\alias{rename_audiomoth_files}
\title{Rename audiomoth files to be compatible with Arbimon}
\usage{
rename_audiomoth_files(dir_path)
}
\arguments{
\item{dir_path}{a path to the directory of audiomoth files}
}
\value{
a directory with audiomoth files that do not end in 'T'
}
\description{
This function removes the 'T' from the end of audiomoth filenames allowing them to be uploaded to Arbimon
}
\examples{
\dontrun{
rename_audiomoth_files("user/audiomoth_files")
}
}
