% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_header_info.R
\name{extract_header_info}
\alias{extract_header_info}
\title{Extract information from audiomoth wav file headers}
\usage{
extract_header_info(dir_path)
}
\arguments{
\item{dir_path}{path to directory that holds audiomoth files}
}
\value{
dataframe with recording time, settings, battery and temperature
}
\description{
Extract information from audiomoth wav file headers
}
\examples{
extract_header_info(system.file("extdata", package = "audiomoth.tools"))
}
